+++ srcs/binary_tree/bst_print.c
@@ -54,68 +54,3 @@ void	print_bst_pipe(t_pipe *node, int space)
 	if (node->type.type == N_PIPE)
 		printf("[PIPE]\n");
 }
-
-void	print_bst_exec(t_exec *node, int space)
-{
-	int	i;
-	int	j;
-
-	i = 0;
-	j = 0;
-	if (node->outfiles)
-	{
-		i = 0;
-		// Add indentation before printing infiles
-		while (i < space + 2)
-		{
-			printf(" ");
-			i++;
-		}
-		printf("[OUTFILES]: ");
-		for (i = 0; node->outfiles[i]; i++)
-		{
-			printf("%s", node->outfiles[i]);
-			if (node->outfiles[i + 1])
-				printf(", "); // Separate infiles with commas
-		}
-	}
-	printf("\n");
-	while (i < space)
-	{
-		printf(" ");
-		i++;
-	}
-	printf("[EXEC]\n");
-	if (node->argv)
-	{
-		while (node->argv[j])
-		{
-			i = 0;
-			while (i < space + 2)
-			{
-				printf(" ");
-				i++;
-			}
-			printf("[%s]\n", node->argv[j]);
-			j++;
-		}
-	}
-	if (node->infiles)
-	{
-		i = 0;
-		// Add indentation before printing infiles
-		while (i < space + 2)
-		{
-			printf(" ");
-			i++;
-		}
-		printf("[INFILES]: ");
-		for (i = 0; node->infiles[i]; i++)
-		{
-			printf("%s", node->infiles[i]);
-			if (node->infiles[i + 1])
-				printf(", "); // Separate infiles with commas
-		}
-		printf("\n");
-	}
-}
