+++ srcs/terminal/terminal.c
@@ -50,6 +50,7 @@ void	terminal(t_shell *shell, char **envp)
 	shell->envp_arr = env_arr(shell);
 	shell->path = path_list(shell, envp);
 	shell->root = build_tree(shell, shell->token_lst);
+	print_bst(shell->root, 5);
 	if (fork() == 0)
 		exec_tree(shell, shell->root);
 	waitpid(-1, &status, 0);
